Memory optimizaations

+++ config.hpp
@@ -27,8 +27,8 @@
   //#define SERIAL_PLOT           //Activates Serial Plotter output
   #define DEBUG_DISP            //Activates Debug output via Display
 
-  const int T_SHOWDEBUG = 1000;
-  const int AVERAGE_OF  = 5;     //how many readings will be averaged per measurement.
+  const uint16_t T_SHOWDEBUG = 1000;
+  const uint8_t AVERAGE_OF  = 5;     //how many readings will be averaged per measurement.
 
   /*---------------------------------------------------------------*/
   #define NUM_SENSORS        3         //Amount of sensors connected.
@@ -38,39 +38,39 @@
 
   //Sensor 1
   #define SENSOR_1_ID             F("Monstera") //A Name by which this sensor is identified (e.g. Name of the corresponding plant)
-  const int SENSOR_1_PIN =        A0;           //Analog Pin to which the sensor is connected.
-  const int SENSOR_1_PPIN =       11;           //Pin that powers the sensor (if applicable). Set to -1 if not used.
+  const uint8_t SENSOR_1_PIN =        A0;           //Analog Pin to which the sensor is connected.
+  const uint8_t SENSOR_1_PPIN =       11;           //Pin that powers the sensor (if applicable). Set to -1 if not used.
   //Sensor 2
   #define SENSOR_2_ID             F("Schaefl.")  //A Name by which this sensor is identified (e.g. Name of the corresponding plant)
-  const int SENSOR_2_PIN =        A1;           //Analog Pin to which the sensor is connected.
-  const int SENSOR_2_PPIN =       12;           //Pin that powers the sensor (if applicable). Set to -1 if not used.
+  const uint8_t SENSOR_2_PIN =        A1;           //Analog Pin to which the sensor is connected.
+  const uint8_t SENSOR_2_PPIN =       12;           //Pin that powers the sensor (if applicable). Set to -1 if not used.
 
   //Sensor 3
   #define SENSOR_3_ID             F("Gl.Feder") //A Name by which this sensor is identified (e.g. Name of the corresponding plant)
-  const int SENSOR_3_PIN =        A2;           //Analog Pin to which the sensor is connected.
-  const int SENSOR_3_PPIN =       13;           //Pin that powers the sensor (if applicable). Set to -1 if not used.
+  const uint8_t SENSOR_3_PIN =        A2;           //Analog Pin to which the sensor is connected.
+  const uint8_t SENSOR_3_PPIN =       13;           //Pin that powers the sensor (if applicable). Set to -1 if not used.
 
   //Sensor 4
   #define SENSOR_4_ID             F("Monstera") //A Name by which this sensor is identified (e.g. Name of the corresponding plant)
-  const int SENSOR_4_PIN =        A3;           //Analog Pin to which the sensor is connected.
-  const int SENSOR_4_PPIN =       14;           //Pin that powers the sensor (if applicable). Set to -1 if not used.  
+  const uint8_t SENSOR_4_PIN =        A3;           //Analog Pin to which the sensor is connected.
+  const uint8_t SENSOR_4_PPIN =       14;           //Pin that powers the sensor (if applicable). Set to -1 if not used.  
 
   //Sensor 5
   #define SENSOR_5_ID             F("Monstera") //A Name by which this sensor is identified (e.g. Name of the corresponding plant)
-  const int SENSOR_5_PIN =        A6;           //Analog Pin to which the sensor is connected.
-  const int SENSOR_5_PPIN =       15;           //Pin that powers the sensor (if applicable). Set to -1 if not used.
+  const uint8_t SENSOR_5_PIN =        A6;           //Analog Pin to which the sensor is connected.
+  const uint8_t SENSOR_5_PPIN =       15;           //Pin that powers the sensor (if applicable). Set to -1 if not used.
 
   //Sensor 6
   #define SENSOR_6_ID             F("Monstera") //A Name by which this sensor is identified (e.g. Name of the corresponding plant)
-  const int SENSOR_6_PIN =        A7;           //Analog Pin to which the sensor is connected.
-  const int SENSOR_6_PPIN =       16;           //Pin that powers the sensor (if applicable). Set to -1 if not used.
+  const uint8_t SENSOR_6_PIN =        A7;           //Analog Pin to which the sensor is connected.
+  const uint8_t SENSOR_6_PPIN =       16;           //Pin that powers the sensor (if applicable). Set to -1 if not used.
 
   //Calibration values for the sensors.
   //These values are used to map the analog reading to a percentage value.
   //You need to measure these values for your specific sensor and setup.
   
-  const int SENSOR_CALIBRATED_MIN = 360; //meassured analog value if sensor is in water.
-  const int SENSOR_CALIBRATED_MAX = 790; //meassured analog value if sensor is dry.
+  const uint16_t SENSOR_CALIBRATED_MIN = 360; //meassured analog value if sensor is in water.
+  const uint16_t SENSOR_CALIBRATED_MAX = 790; //meassured analog value if sensor is dry.
 
 
 
@@ -81,11 +81,11 @@
 /************************************************************************/
 
 //PIN Config.
-  const int MAX_SENSORS = 6;      //Maximum amount of sensors supported by this software.
+  const uint8_t MAX_SENSORS = 6;      //Maximum amount of sensors supported by this software.
   static_assert(NUM_SENSORS <= MAX_SENSORS, "Error: NUM_SENSORS exceeds MAX_SENSORS. Please adjust configuration.");
 
 //Advanced Config.
-  const int SERIAL_TRANSMIT_DELAY = 5; //Delay in ms between sending messages via Serial. Increase this if you experience problems with your serial connection.
+  const uint8_t SERIAL_TRANSMIT_DELAY = 5; //Delay in ms between sending messages via Serial. Increase this if you experience problems with your serial connection.
   #define BAUDRATE          57600    //Boudrate for serial connection, match this to your terminal. If you don't know what this means, don't change it.
   #define ANALOG_REF        DEFAULT  //Configures the reference voltage used for analog input.
 #endif
